refactor: processor `css-custom-property` feature extract
- moved all code related to feature file
- new `analyzeDeclaration` feature hook
