+++ packages/core/src/stylable-processor.ts
@@ -1,14 +1,14 @@
 import path from 'path';
 import * as postcss from 'postcss';
-import postcssValueParser from 'postcss-value-parser';
 import { deprecatedStFunctions } from './custom-values';
 import { Diagnostics } from './diagnostics';
 import { parseSelector as deprecatedParseSelector } from './deprecated/deprecated-selector-utils';
 import { murmurhash3_32_gc } from './murmurhash';
 import { knownPseudoClassesWithNestedSelectors } from './native-reserved-lists';
 import { StylableMeta } from './stylable-meta';
-import type {
+import {
     ClassSymbol,
+    CSSCustomProperty,
     CSSVarSymbol,
     ElementSymbol,
     RefedMixin,
@@ -40,12 +40,10 @@ import {
     rootValueMapping,
     SBTypesParsers,
     stValuesMap,
-    validateAllowedNodesUntil,
     valueMapping,
 } from './stylable-value-parsers';
-import { deprecated, filename2varname, globalValue, stripQuotation } from './utils';
+import { deprecated, filename2varname, stripQuotation } from './utils';
 import { ignoreDeprecationWarn } from './helpers/deprecation';
-import { validateAtProperty } from './helpers/css-custom-property';
 
 const parseStates = SBTypesParsers[valueMapping.states];
 const parseGlobal = SBTypesParsers[valueMapping.global];
@@ -97,12 +95,6 @@ export const processorWarnings = {
     GLOBAL_CSS_VAR_MISSING_COMMA(name: string) {
         return `"@st-global-custom-property" received the value "${name}", but its values must be comma separated`;
     },
-    ILLEGAL_CSS_VAR_USE(name: string) {
-        return `a custom css property must begin with "--" (double-dash), but received "${name}"`;
-    },
-    ILLEGAL_CSS_VAR_ARGS(name: string) {
-        return `custom property "${name}" usage (var()) must receive comma separated values`;
-    },
     INVALID_NAMESPACE_REFERENCE() {
         return 'st-namespace-reference dose not have any value';
     },
@@ -153,13 +145,8 @@ export class StylableProcessor implements FeatureContext {
         root.walkDecls((decl) => {
             if (stValuesMap[decl.prop]) {
                 this.handleDirectives(decl.parent as SRule, decl);
-            } else if (isCSSVarProp(decl.prop)) {
-                this.addCSSVarDefinition(decl);
             }
-
-            if (decl.value.includes('var(')) {
-                this.handleCSSVarUse(decl);
-            }
+            CSSCustomProperty.hooks.analyzeDeclaration({ context: this, decl });
 
             this.collectUrls(decl);
         });
@@ -238,8 +225,7 @@ export class StylableProcessor implements FeatureContext {
                     this.meta.scopes.push(atRule);
                     break;
                 case 'property': {
-                    this.addCSSVarDefinition(atRule);
-                    validateAtProperty(atRule, this.diagnostics);
+                    CSSCustomProperty.hooks.analyzeAtRule({ context: this, atRule });
                     break;
                 }
                 case 'st-global-custom-property': {
@@ -516,73 +502,6 @@ export class StylableProcessor implements FeatureContext {
             });
         });
         rule.remove();
-    }
-
-    protected handleCSSVarUse(decl: postcss.Declaration) {
-        const parsed = postcssValueParser(decl.value);
-        parsed.walk((node) => {
-            if (node.type === 'function' && node.value === 'var' && node.nodes) {
-                const varName = node.nodes[0];
-                if (!validateAllowedNodesUntil(node, 1)) {
-                    const args = postcssValueParser.stringify(node.nodes);
-                    this.diagnostics.warn(decl, processorWarnings.ILLEGAL_CSS_VAR_ARGS(args), {
-                        word: args,
-                    });
-                }
-
-                this.addCSSVar(postcssValueParser.stringify(varName).trim(), decl, false);
-            }
-        });
-    }
-
-    protected addCSSVarDefinition(node: postcss.Declaration | postcss.AtRule) {
-        let varName = node.type === 'atrule' ? node.params.trim() : node.prop.trim();
-        let isGlobal = false;
-
-        const globalVarName = globalValue(varName);
-
-        if (globalVarName !== undefined) {
-            varName = globalVarName.trim();
-            isGlobal = true;
-        }
-
-        if (node.type === 'atrule' && STSymbol.get(this.meta, varName)) {
-            this.diagnostics.warn(node, STSymbol.diagnostics.REDECLARE_SYMBOL(varName), {
-                word: varName,
-            });
-        }
-
-        this.addCSSVar(varName, node, isGlobal);
-    }
-
-    protected addCSSVar(
-        varName: string,
-        node: postcss.Declaration | postcss.AtRule,
-        global: boolean
-    ) {
-        if (isCSSVarProp(varName)) {
-            if (!this.meta.cssVars[varName]) {
-                const cssVarSymbol: CSSVarSymbol = {
-                    _kind: 'cssVar',
-                    name: varName,
-                    global,
-                };
-                this.meta.cssVars[varName] = cssVarSymbol;
-                const prevSymbol = STSymbol.get(this.meta, varName);
-                const override = node.type === `atrule` || !prevSymbol;
-                if (override) {
-                    STSymbol.addSymbol({
-                        context: this,
-                        symbol: cssVarSymbol,
-                        safeRedeclare: true,
-                    });
-                }
-            }
-        } else {
-            this.diagnostics.warn(node, processorWarnings.ILLEGAL_CSS_VAR_USE(varName), {
-                word: varName,
-            });
-        }
     }
 
     protected handleDirectives(rule: SRule, decl: postcss.Declaration) {
